+++ example/lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:current/current.dart';
+import 'package:current_counter_example/space_mission_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:uuid/uuid.dart';
 
 import 'application_view_model.dart';
 import 'mission_control_shell.dart';
-import 'mission_control_theme.dart';
 
 void main() {
   runApp(MyApp());
