+++ example/lib/mission_section.dart
@@ -1,0 +1,7 @@
+enum MissionSection {
+  overview,
+  telemetry,
+  flightForms,
+  starMap,
+  launchEvents,
+}
