+++ example/lib/pages/star_map_page.dart
@@ -1,4 +1,5 @@
 import 'package:current/current.dart';
+import 'package:current_counter_example/space_mission_theme.dart';
 import 'package:flutter/material.dart';
 
 import '../mission_control_theme.dart';
